Adjust detection of default logo
In production style builds, the logo has a hash at the end of the
filename, so the match can't be as specific as it was.

This shouldn't match anything else since it is just on imported and
archived challenges.

If the default logo file is changed or renamed, this would need to be
updated.