CHAL-1245 Solvers Internal Error
Solvers are receiving an internal server error when attempting to request reactivation.
The issue lies in teh first and last name being required on a user struct. These are passed as nil
for most solvers as they were never named. This commit ensures users are assigned empty strings
as names if they are currently nil.